Add unit test for encodeChunkFileName
